+++ src/vstarstack/library/fine_shift/image_wave.h
@@ -1,8 +1,8 @@
 
 /* Interpolation methods */
-double image_wave_interpolation_1d(double fm1, double f0, double f1, double f2, double x);
+double image_grid_interpolation_1d(double fm1, double f0, double f1, double f2, double x);
 
-double image_wave_interpolation_2d(double fm1m1, double f0m1, double f1m1, double f2m1,
+double image_grid_interpolation_2d(double fm1m1, double f0m1, double f1m1, double f2m1,
                         double fm10,  double f00,  double f10,  double f20,
                         double fm11,  double f01,  double f11,  double f21,
                         double fm12,  double f02,  double f12,  double f22,
